+++ x/crosschain/keeper/cctx_utils_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"fmt"
 	"math/big"
 	"testing"
 
@@ -29,7 +30,8 @@ func TestGetRevertGasLimit(t *testing.T) {
 		gasLimit, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
 				CoinType: common.CoinType_Zeta,
-			}})
+			},
+		})
 		require.NoError(t, err)
 		require.Equal(t, uint64(0), gasLimit)
 	})
@@ -46,9 +48,10 @@ func TestGetRevertGasLimit(t *testing.T) {
 		require.NoError(t, err)
 
 		gasLimit, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
+				CoinType:      common.CoinType_Gas,
 			}})
 		require.NoError(t, err)
 		require.Equal(t, uint64(42), gasLimit)
@@ -76,10 +79,11 @@ func TestGetRevertGasLimit(t *testing.T) {
 		require.NoError(t, err)
 
 		gasLimit, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: chainID,
 				Asset:         asset,
+				CoinType:      common.CoinType_ERC20,
 			}})
 		require.NoError(t, err)
 		require.Equal(t, uint64(42), gasLimit)
@@ -89,9 +93,10 @@ func TestGetRevertGasLimit(t *testing.T) {
 		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
 
 		_, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: 999999,
+				CoinType:      common.CoinType_Gas,
 			}})
 		require.ErrorIs(t, err, types.ErrForeignCoinNotFound)
 	})
@@ -110,9 +115,10 @@ func TestGetRevertGasLimit(t *testing.T) {
 
 		// no contract deployed therefore will fail
 		_, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
+				CoinType:      common.CoinType_Gas,
 			}})
 		require.ErrorIs(t, err, fungibletypes.ErrContractCall)
 	})
@@ -121,9 +127,10 @@ func TestGetRevertGasLimit(t *testing.T) {
 		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
 
 		_, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: 999999,
+				CoinType:      common.CoinType_ERC20,
 			}})
 		require.ErrorIs(t, err, types.ErrForeignCoinNotFound)
 	})
@@ -145,9 +152,9 @@ func TestGetRevertGasLimit(t *testing.T) {
 		// no contract deployed therefore will fail
 		_, err := k.GetRevertGasLimit(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: chainID,
 				Asset:         asset,
+				CoinType:      common.CoinType_ERC20,
 			}})
 		require.ErrorIs(t, err, fungibletypes.ErrContractCall)
 	})
@@ -190,4 +197,23 @@ func TestGetAbortedAmount(t *testing.T) {
 		a := crosschainkeeper.GetAbortedAmount(cctx)
 		require.Equal(t, sdkmath.ZeroUint(), a)
 	})
+}
+
+func Test_IsPending(t *testing.T) {
+	tt := []struct {
+		status   types.CctxStatus
+		expected bool
+	}{
+		{types.CctxStatus_PendingInbound, false},
+		{types.CctxStatus_PendingOutbound, true},
+		{types.CctxStatus_PendingRevert, true},
+		{types.CctxStatus_Reverted, false},
+		{types.CctxStatus_Aborted, false},
+		{types.CctxStatus_OutboundMined, false},
+	}
+	for _, tc := range tt {
+		t.Run(fmt.Sprintf("status %s", tc.status), func(t *testing.T) {
+			require.Equal(t, tc.expected, crosschainkeeper.IsPending(types.CrossChainTx{CctxStatus: &types.Status{Status: tc.status}}))
+		})
+	}
 }