+++ x/crosschain/keeper/refund_test.go
@@ -26,11 +26,11 @@ func TestKeeper_RefundAmountOnZetaChainGas(t *testing.T) {
 
 		err := k.RefundAmountOnZetaChainGas(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.NewUint(20),
+				CoinType:      common.CoinType_Gas,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.NewUint(42),
@@ -53,11 +53,11 @@ func TestKeeper_RefundAmountOnZetaChainGas(t *testing.T) {
 
 		err := k.RefundAmountOnZetaChainGas(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.NewUint(20),
+				CoinType:      common.CoinType_Gas,
 			},
 		},
 			sender,
@@ -75,11 +75,11 @@ func TestKeeper_RefundAmountOnZetaChainGas(t *testing.T) {
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		err := k.RefundAmountOnZetaChainGas(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.NewUint(20),
+				CoinType:      common.CoinType_Gas,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.NewUint(42),
@@ -100,11 +100,11 @@ func TestKeeper_RefundAmountOnZetaChainGas(t *testing.T) {
 
 		err := k.RefundAmountOnZetaChainGas(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.ZeroUint(),
+				CoinType:      common.CoinType_Gas,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.ZeroUint(),
@@ -126,11 +126,11 @@ func TestKeeper_RefundAmountOnZetaChainZeta(t *testing.T) {
 
 		err := k.RefundAmountOnZetaChainZeta(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.NewUint(20),
+				CoinType:      common.CoinType_Gas,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.NewUint(42),
@@ -139,7 +139,7 @@ func TestKeeper_RefundAmountOnZetaChainZeta(t *testing.T) {
 			sender,
 		)
 		require.NoError(t, err)
-		coin := sdkk.BankKeeper.GetBalance(ctx, sdk.AccAddress(sender.Bytes()), config.BaseDenom)
+		coin := sdkk.BankKeeper.GetBalance(ctx, sender.Bytes(), config.BaseDenom)
 		fmt.Println(coin.Amount.String())
 		require.Equal(t, "42", coin.Amount.String())
 	})
@@ -151,11 +151,11 @@ func TestKeeper_RefundAmountOnZetaChainZeta(t *testing.T) {
 
 		err := k.RefundAmountOnZetaChainZeta(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.NewUint(20),
+				CoinType:      common.CoinType_Gas,
 			},
 		},
 			sender,
@@ -171,12 +171,13 @@ func TestKeeper_RefundAmountOnZetaChainZeta(t *testing.T) {
 		chainID := getValidEthChainID(t)
 
 		err := k.RefundAmountOnZetaChainZeta(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_Gas,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				TxOrigin:      sender.String(),
 				Amount:        math.ZeroUint(),
+				CoinType:      common.CoinType_Gas,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.ZeroUint(),
@@ -210,12 +211,13 @@ func TestKeeper_RefundAmountOnZetaChainERC20(t *testing.T) {
 		)
 
 		err := k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				Asset:         asset,
 				Amount:        math.NewUint(42),
+				CoinType:      common.CoinType_ERC20,
 			},
 			OutboundTxParams: []*types.OutboundTxParams{{
 				Amount: math.NewUint(42),
@@ -232,12 +234,13 @@ func TestKeeper_RefundAmountOnZetaChainERC20(t *testing.T) {
 
 		// can refund again
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: chainID,
 				Sender:        sender.String(),
 				Asset:         asset,
 				Amount:        math.NewUint(42),
+				CoinType:      common.CoinType_ERC20,
 			}},
 			sender,
 		)
@@ -251,10 +254,12 @@ func TestKeeper_RefundAmountOnZetaChainERC20(t *testing.T) {
 		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
 
 		err := k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType: common.CoinType_Zeta,
 				Amount:   math.NewUint(42),
+				CoinType: common.CoinType_Zeta,
 			}},
+
 			sample.EthAddress(),
 		)
 		require.ErrorContains(t, err, "unsupported coin type")
@@ -262,38 +267,42 @@ func TestKeeper_RefundAmountOnZetaChainERC20(t *testing.T) {
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
 			InboundTxParams: &types.InboundTxParams{
 				CoinType: common.CoinType_Gas,
-			}},
+			},
+		},
 			sample.EthAddress(),
 		)
 		require.ErrorContains(t, err, "unsupported coin type")
 
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: 999999,
 				Amount:        math.NewUint(42),
+				CoinType:      common.CoinType_ERC20,
 			}},
 			sample.EthAddress(),
 		)
 		require.ErrorContains(t, err, "only EVM chains are supported")
 
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: getValidEthChainID(t),
 				Sender:        sample.EthAddress().String(),
 				Amount:        math.Uint{},
+				CoinType:      common.CoinType_ERC20,
 			}},
 			sample.EthAddress(),
 		)
 		require.ErrorContains(t, err, "no amount to refund")
 
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: getValidEthChainID(t),
 				Sender:        sample.EthAddress().String(),
 				Amount:        math.ZeroUint(),
+				CoinType:      common.CoinType_ERC20,
 			}},
 			sample.EthAddress(),
 		)
@@ -301,12 +310,13 @@ func TestKeeper_RefundAmountOnZetaChainERC20(t *testing.T) {
 
 		// the foreign coin has not been set
 		err = k.RefundAmountOnZetaChainERC20(ctx, types.CrossChainTx{
+
 			InboundTxParams: &types.InboundTxParams{
-				CoinType:      common.CoinType_ERC20,
 				SenderChainId: getValidEthChainID(t),
 				Sender:        sample.EthAddress().String(),
 				Asset:         sample.EthAddress().String(),
 				Amount:        math.NewUint(42),
+				CoinType:      common.CoinType_ERC20,
 			}},
 			sample.EthAddress(),
 		)