+++ x/crosschain/types/message_abort_stuck_cctx.go
@@ -43,8 +43,8 @@ func (msg *MsgAbortStuckCCTX) ValidateBasic() error {
 	if err != nil {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
 	}
-	if len(msg.CctxIndex) != 66 {
-		return ErrInvalidCCTXIndex
+	if len(msg.CctxIndex) != ZetaIndexLength {
+		return ErrInvalidIndexValue
 	}
 	return nil
 }