+++ changelog.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Refactor
+
+* [1853](https://github.com/zeta-chain/node/pull/1853) - refactor vote inbound tx and vote outbound tx
+
+
 ### Breaking Changes
 
 * Admin policies have been moved from `observer` to a new module `authority`.